python-consumer/requirements.txt
kafka-python
websockets
python-socketio[client]